Move to config instead
